Squashed commit of the following:
commit cbc4af57f025ad41d53238459a593c11def22699
Author: Lvv.me <cntrump@gmail.com>
Date:   Sun Jul 4 23:12:21 2021 +0800

    Remove `f` suffix

commit 720d550a9078314f1abbfc446dc3cc48414965dd
Author: Lvv.me <cntrump@gmail.com>
Date:   Sun Jul 4 22:09:00 2021 +0800

    Replace float with double

commit 460928ec85e498768a65c0140230005e762e61f2
Author: Lvv.me <cntrump@gmail.com>
Date:   Sun Jul 4 18:08:17 2021 +0800

    use offical yoga

+++ tests/CompactValueTest.cpp
@@ -20,7 +20,7 @@ const auto tooLargePercent =
     nextafter(CompactValue::UPPER_BOUND_PERCENT, INFINITY);
 
 TEST(YogaTest, compact_value_can_represent_undefined) {
-  auto c = CompactValue{YGValue{12.5f, YGUnitUndefined}};
+  auto c = CompactValue{YGValue{12.5, YGUnitUndefined}};
   YGValue v = c;
   ASSERT_EQ(v, YGValueUndefined);
   ASSERT_NE(v, YGValueAuto);
@@ -32,7 +32,7 @@ TEST(YogaTest, compact_value_can_represent_undefined) {
 
 TEST(YogaTest, compact_value_manages_infinity_as_undefined) {
   auto c = CompactValue{
-      YGValue{std::numeric_limits<float>::infinity(), YGUnitUndefined}};
+      YGValue{std::numeric_limits<double>::infinity(), YGUnitUndefined}};
   YGValue v = c;
   ASSERT_EQ(v, YGValueUndefined);
   ASSERT_NE(v, YGValueAuto);
@@ -299,23 +299,23 @@ TEST(YogaTest, dedicated_unit_factories) {
   ASSERT_EQ(CompactValue::ofUndefined(), CompactValue(YGValueUndefined));
   ASSERT_EQ(CompactValue::ofAuto(), CompactValue(YGValueAuto));
   ASSERT_EQ(
-      CompactValue::of<YGUnitPoint>(-9876.5f),
-      CompactValue(YGValue{-9876.5f, YGUnitPoint}));
+      CompactValue::of<YGUnitPoint>(-9876.5),
+      CompactValue(YGValue{-9876.5, YGUnitPoint}));
   ASSERT_EQ(
-      CompactValue::of<YGUnitPercent>(123.456f),
-      CompactValue(YGValue{123.456f, YGUnitPercent}));
+      CompactValue::of<YGUnitPercent>(123.456),
+      CompactValue(YGValue{123.456, YGUnitPercent}));
 }
 
 TEST(YogaTest, dedicated_unit_maybe_factories) {
   ASSERT_EQ(
-      CompactValue::ofMaybe<YGUnitPoint>(-9876.5f),
-      CompactValue(YGValue{-9876.5f, YGUnitPoint}));
+      CompactValue::ofMaybe<YGUnitPoint>(-9876.5),
+      CompactValue(YGValue{-9876.5, YGUnitPoint}));
   ASSERT_EQ(
       CompactValue::ofMaybe<YGUnitPoint>(YGUndefined),
       CompactValue(YGValueUndefined));
   ASSERT_EQ(
-      CompactValue::ofMaybe<YGUnitPercent>(123.456f),
-      CompactValue(YGValue{123.456f, YGUnitPercent}));
+      CompactValue::ofMaybe<YGUnitPercent>(123.456),
+      CompactValue(YGValue{123.456, YGUnitPercent}));
   ASSERT_EQ(
       CompactValue::ofMaybe<YGUnitPercent>(YGUndefined),
       CompactValue(YGValueUndefined));
@@ -324,7 +324,7 @@ TEST(YogaTest, dedicated_unit_maybe_factories) {
 TEST(YogaTest, can_be_assigned_from_YGValue) {
   CompactValue c{};
 
-  YGValue v{2.0f, YGUnitPercent};
+  YGValue v{2.0, YGUnitPercent};
   c = v;
   ASSERT_EQ((YGValue) c, v);
